+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 ## Changelog
 
+### 1.1.0 (January 8, 2016)
+
+- `ADDED` 3DM model parameters settings ( label, tooltip, lightLevel, emitRedstone, buttonMode, collidable )
+- `ADDED` Tints ( object color ) export ( material diffuse color is used as source )
+
 ### 1.0.0 (January 7, 2016)
 
 - Initial release